requirements-optional.txt
tiktoken>=0.3.2 # openai calculate token

#voice
pydub>=0.25.1 # need ffmpeg
SpeechRecognition # google speech to text
gTTS>=2.3.1 # google text to speech
pyttsx3>=2.90 # pytsx text to speech
baidu_aip>=4.16.10 # baidu voice
azure-cognitiveservices-speech # azure voice
numpy<=1.24.2
langid # language detect

#install plugin
dulwich

# wechaty
wechaty>=0.10.7
wechaty_puppet>=0.4.23
pysilk_mod>=1.6.0 # needed by send voice

# wechatmp wechatcom
web.py
wechatpy

# chatgpt-tool-hub plugin

--extra-index-url https://pypi.python.org/simple
chatgpt_tool_hub==0.4.6

# rocketmq
rocketmq-client-python>=2.0.0